requirements.txt
Django==1.7.7
MySQL-python==1.2.5
argparse==1.2.1
boto==2.38.0
pusher==0.8
requests==2.6.0
wsgiref==0.1.2
django-push-notifications==1.2.1
django-uuidfield==0.5.0
